Polish auth screen and fix colour overflow in cloud sync
- Redesign the login/sign-up screen: gradient backdrop, elevated brand
  card, rounded fields, and Google/Apple continue buttons that match the
  app's design language.
- Carry a note's packed ARGB colour as a Long over the wire. Postgres
  stores it unsigned (e.g. 0xFFB8E0F5 -> 4290312015), which overflowed
  the signed Int and failed every pull; toInt() recovers the exact
  signed colour, so the round-trip stays lossless.

+++ core/data/src/commonMain/kotlin/com/androidpoet/materialnotes/data/sync/RemoteNote.kt
@@ -13,7 +13,11 @@ data class RemoteNote(
     val id: String,
     val title: String,
     val date: String,
-    val backround: Int,
+    // Packed ARGB colour. The local model keeps it as a signed Int (Compose's Color.toArgb()), but
+    // Postgres stores the value unsigned (e.g. 0xFFB8E0F5 → 4290312015), which overflows a signed
+    // Int on the way back. Carry it as a Long over the wire; `.toInt()` truncates the low 32 bits to
+    // recover the exact signed colour, so the round-trip is lossless either way.
+    val backround: Long,
     val content: String,
     @SerialName("created_at") val createdAt: Long,
 )
@@ -22,7 +26,7 @@ fun Note.toRemote(): RemoteNote = RemoteNote(
     id = id,
     title = title,
     date = date,
-    backround = backround,
+    backround = backround.toLong(),
     content = content,
     createdAt = createdAt,
 )
@@ -31,7 +35,7 @@ fun RemoteNote.toDomain(): Note = Note(
     id = id,
     title = title,
     date = date,
-    backround = backround,
+    backround = backround.toInt(),
     content = content,
     createdAt = createdAt,
 )
